modified client and server and packet code

+++ packet.h
@@ -7,7 +7,7 @@
 */
 typedef struct {
 	short opcode;
-	short sequenceNumber;
+	unsigned int sequenceNumber;
 } Header;
 
 typedef struct {
